Proper use of io manager
Behaviour of IO manager depends on platform.  On Unix its noop, as RTS
has an io manager already; on Windows: it starts a thread which fulfills
aysynchronous IO request (throuh IO completion port).
`withIOManager` is used when an application starts and it must be called
only once, so the best place to start is at the very begining of `main`.